+++ contracts/interfaces/IRewardsDistributor.sol
@@ -54,7 +54,7 @@ interface IRewardsDistributor {
 
     function getStrategyRewards(address _strategy) external returns (uint96);
 
-    function sendTokensToContributor(address _to, uint96 _amount) external;
+    function sendTokensToContributor(address _to, uint256 _amount) external;
 
     function getRewards(
         address _garden,
